Add Exception integration & service test, put User tests in seperated classes
