+++ backend/src/test/java/com/github/lukaskraus/backend/service/UserIntegrationTest.java
@@ -1,0 +1,38 @@
+package com.github.lukaskraus.backend.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+class UserIntegrationTest {
+
+    @Autowired
+    MockMvc mockMvc;
+
+    @DirtiesContext
+    @WithMockUser(username = "kevin")
+    @Test
+    void showUsernameKevin_WhenLoggedIn() throws Exception {
+        mockMvc.perform(get("/api/users/me"))
+                .andExpect(status().isOk())
+                .andExpect(content().string("kevin"));
+    }
+
+    @DirtiesContext
+    @Test
+    void showAnonymousUser_WhenNotLoggedIn() throws Exception {
+        mockMvc.perform(get("/api/users/me"))
+                .andExpect(status().isOk())
+                .andExpect(content().string("anonymousUser"));
+    }
+}
